+++ src/Native/ArmNative/Arm64Disassembler.cpp
@@ -74,7 +74,9 @@ INativeInstruction * Arm64Disassembler::NextInstruction()
 	else
 	{
 		auto info = NativeInstructionInfo{
-			uAddr, 4, static_cast<uint32_t>(InstructionClass::Linear), instr->id
+			uAddr, 4,
+			static_cast<uint32_t>(InstructionClass::Linear),
+			static_cast<int32_t>(instr->id)
 		};
 		return new Arm64Instruction(instr, info);
 	}
