+++ src/Native/ArmNative/ThumbDisassembler.cpp
@@ -72,7 +72,9 @@ INativeInstruction * ThumbDisassembler::NextInstruction()
 	else
 	{
 		auto info = NativeInstructionInfo{
-			uAddr, 4, static_cast<uint32_t>(InstructionClass::Linear), instr->id
+			uAddr, 4,
+			static_cast<uint32_t>(InstructionClass::Linear), 
+			static_cast<int32_t>(instr->id)
 		};
 		return new NativeInstruction(instr, info);
 	}
